+++ src/pages/ListagemUsuarios/index.js
@@ -1,0 +1,3 @@
+import ListagemUsuarios from './ListagemUsuarios';
+
+export default ListagemUsuarios;
